SequenceDiagram
The actor makes an attempt to take the quiz using the UI of the software. Then the received
request is responded with the list of questions received from the database.
The software then sends a message to respond to the questions to the actor.
Once the complete answers are received, the software sends the answer to the database where
they are stored and forwarded to the ML model for regression analysis.
Once the analysis is complete, the ML model provides a study path based on the analysis of the
answers to the questions. The study path is stored in the database and sent forward to the
software where it is displayed to the actor.
In case the questionnaire is not completed, the software responds with a error message to the
actor